# meta: optimize typesmap Iterate for common cases