@YuriyNasretdinov 2019-04-03
Was is that slow that it affected overall performance? You can avoid allocations another way probably :)