@YuriyNasretdinov 2019-04-08
I suppose that if you can just make sure that the string slice is always allocated on stack, this will give pretty good performance as well. But I am not sure how to do it :).